requirements.txt
streamlit
httpx-oauth
pyjwt